+++ src/main/java/net/oneandone/pommes/cli/DatabaseAdd.java
@@ -22,7 +22,9 @@ import net.oneandone.pommes.model.Field;
 import net.oneandone.pommes.model.Pom;
 import net.oneandone.pommes.project.Project;
 import net.oneandone.pommes.repository.Repository;
+import net.oneandone.pommes.scm.Scm;
 import net.oneandone.sushi.fs.NodeInstantiationException;
+import net.oneandone.sushi.fs.file.FileNode;
 import org.apache.lucene.document.Document;
 
 import java.io.IOException;
@@ -214,6 +216,8 @@ public class DatabaseAdd extends Base {
             Project project;
             Pom pom;
             String fixed;
+            FileNode checkout;
+            Scm scm;
 
             while (true) {
                 project = src.take();
@@ -229,6 +233,13 @@ public class DatabaseAdd extends Base {
                             throw new IllegalStateException(fixed);
                         }
                         fixed = fixed.substring(0, fixed.lastIndexOf('/'));
+                        if (fixed.startsWith("file:")) {
+                            checkout = (FileNode) environment.world().node(fixed);
+                            scm = Scm.probeCheckout(checkout);
+                            if (scm != null) {
+                                fixed = scm.getUrl(checkout);
+                            }
+                        }
                         if (!fixed.equals(pom.scm)) {
                             environment.console().info.println("WARNING: fixing scm " + pom.scm + " -> " + fixed);
                             pom = pom.clone(fixed);